refactor: move nested PM to its own file
